Use fixed sized integers
Let's use the real integer size.

Since we are emulating an 8-bit CPU on a usually bigger one, we just let
the compiler handle the necessary optimisations if integer promotion is
useful.

But then we also let the compiler ensure that there's no overflow or
wrong type size. This is specially important when we do integer
arithmetics